fix(state): align issue #27 with client/ app
