+++ client/src/state/uiPreferences.tsx
@@ -1,44 +1,19 @@
+/* eslint-disable react-refresh/only-export-components */
+
 import {
   createContext,
-  ReactNode,
+  type ReactNode,
   useContext,
   useEffect,
   useMemo,
   useState,
 } from 'react';
-
-export interface UiPreferences {
-  sidebarCollapsed: boolean;
-}
-
-const STORAGE_KEY = 'aiaf.uiPreferences.v1';
-
-const defaultPreferences: UiPreferences = {
-  sidebarCollapsed: false,
-};
-
-export function loadUiPreferences(): UiPreferences {
-  try {
-    const raw = localStorage.getItem(STORAGE_KEY);
-    if (!raw) return defaultPreferences;
-
-    const parsed = JSON.parse(raw) as Partial<UiPreferences>;
-    return {
-      ...defaultPreferences,
-      ...parsed,
-    };
-  } catch {
-    return defaultPreferences;
-  }
-}
-
-export function saveUiPreferences(preferences: UiPreferences): void {
-  try {
-    localStorage.setItem(STORAGE_KEY, JSON.stringify(preferences));
-  } catch {
-    // ignore (storage may be unavailable)
-  }
-}
+import {
+  getDefaultUiPreferences,
+  loadUiPreferences,
+  saveUiPreferences,
+  type UiPreferences,
+} from './uiPreferencesStorage';
 
 interface UiPreferencesContextValue {
   preferences: UiPreferences;
@@ -64,7 +39,7 @@ export function UiPreferencesProvider({ children }: { children: ReactNode }) {
       preferences,
       setSidebarCollapsed: (collapsed: boolean) =>
         setPreferences((prev) => ({ ...prev, sidebarCollapsed: collapsed })),
-      resetPreferences: () => setPreferences(defaultPreferences),
+      resetPreferences: () => setPreferences(getDefaultUiPreferences()),
     }),
     [preferences],
   );
